Fixed compilation error in Scheduler.h #rb trivial
[CL 23966881 by dmytro vovk in ue5-main branch]

+++ Engine/Source/Runtime/Core/Public/Async/Fundamental/Scheduler.h
@@ -50,7 +50,7 @@ namespace LowLevelTasks
 
 		uint32 GetNodeIndex(const NodeType* Node) const
 		{
-			return (Node == nullptr) ? EVENT_INDEX_NONE : (Node - NodesArray.GetData());
+			return (Node == nullptr) ? EVENT_INDEX_NONE : uint32(Node - NodesArray.GetData());
 		}
 
 	public:
